inventory_simulator/requirements.txt
mysql-connector-python
pandas
numpy
python-dotenv